addded openrouter apis for AI, added export data section, calorie intake will now use ai or harcoded or favorite

+++ app/src/main/java/com/dutch/thryve/data/repository/FirebaseRepositoryImpl.kt
@@ -32,7 +32,7 @@ class FirebaseRepositoryImpl @Inject constructor(
                 querySnapshot.documents.mapNotNull { document ->
                     try {
                         val exerciseName = document.getString("exerciseName") ?: ""
-                        val weight = document.getLong("weight")?.toInt() ?: 0
+                        val weight = document.getLong("weight")?.toDouble() ?: 0.0
                         val reps = document.getLong("reps")?.toInt() ?: 0
 
                         val dateObject = document.get("date")
